+++ src/senaite/core/upgrade/v02_06_000.py
@@ -2445,3 +2445,16 @@ def del_raw_analyses(tool):
         sample._p_deactivate()
 
     logger.info("Remove Analyses raw attribute from samples [DONE]")
+
+
+def ensure_valid_sticker_templates(tool):
+    logger.info("Ensure sample types have valid sticker templates ...")
+    query = {"portal_type": "SampleType"}
+    brains = api.search(query, SETUP_CATALOG)
+    total = len(brains)
+    for num, brain in enumerate(brains):
+        logger.info("Ensure sample types have valid sticker templates {0}/{1}"
+                    .format(num+1, total))
+        obj = api.get_object(brain)
+        obj.setAdmittedStickerTemplates(obj.getAdmittedStickerTemplates())
+    logger.info("Ensure sample types have valid sticker templates [DONE]")